Extract VectorizedColumnReader from VectorizedParquetRecordReader
